requirements.txt
accelerate==0.25.0
ai-dive==0.0.1
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
datasets==2.14.5
dill==0.3.7
distlib==0.3.7
filelock==3.12.3
frozenlist==1.4.0
fsspec==2023.6.0
huggingface-hub==0.19.4
idna==3.6
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.2
pandas==2.0.1
Pillow==10.1.0
pipenv==2023.9.7
platformdirs==3.10.0
protobuf==4.25.1
psutil==5.9.6
pyarrow==14.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sympy==1.12
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tokenizers==0.13.3
toml==0.10.2
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.65.0
transformers==4.30.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.24.4
xxhash==3.4.1
yarl==1.9.4